Bug and errors fixing : Adding Error handling if the clubs or competitions data are not in conformity. Adding Condition for If the club or the competition is not found in the database. Adding Error handling if the number of places required entered by the user is not a number. Adding BUG fix handling if the number of club's points is not enough for the number of places required. Adding BUG fix handling if the number of competition's available places is not enough for the number of places required.
Tests :
Adding test if the number of club's points and competition's available places are enough.
Adding test if the number of places required is not a number.
Adding test if the number of club's points is not enough for the number of places required.
Adding test if the number of competition's available places is not enough for the number of places required.